// types/pattern.ts
export enum Pattern {
  NORMAL = 'NORMAL',
  DIAGONAL = 'DIAGONAL',
  FUNKY_DIAGONAL = 'FUNKY_DIAGONAL',
  TRIANGLE = 'TRIANGLE',
  MOTHERBOARD = 'MOTHERBOARD',
  SQUARE = 'SQUARE',
  STRIP = 'STRIP',
  STRIP_Y = 'STRIP_Y',
  CROSS_STRIP = 'CROSS_STRIP',
  SPIRAL = 'SPIRAL',
  BIG_SPIRAL = 'BIG_SPIRAL',
  YING_YANG = 'YING_YANG',
  RADIAL_LINES = 'RADIAL_LINES',
  WAVE = 'WAVE',
  WAVE_X = 'WAVE_X',
  GRADIENT = 'GRADIENT',
  DITHER = 'DITHER',
  HEAT_INSTRUCTION_BYTE_SIZE = 'HEAT_INSTRUCTION_BYTE_SIZE',
  PALINDROME = 'PALINDROME',
  Z = 'Z',
  SQUIGGLY = 'SQUIGGLY'
}